start showing data from an SQLite database
We add an SQLite database that will hold all the persistent data.  For
now it only contains data about users, subject and homework. Code to
deal with the database is isolated in src/data.js.

We also implement a crude day view that uses this data to show all due
homework for a given day.